third_party/WebKit/LayoutTests/svg/custom/selectSubString-expected.txt

PASS svgText.selectSubString(-1, 0) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgText.getSubStringLength(0, -1) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS svgText.getSubStringLength(3, 0) threw exception IndexSizeError: Index or size was negative, or greater than the allowed value..
PASS successfullyParsed is true

TEST COMPLETE
